refactor(desktop): split app.tsx into keepalive/emptystate parts + hooks
