+++ apps/desktop/src/App.tsx
@@ -3,8 +3,8 @@ import { AppShell } from '@goodboy/ui';
 import type { PlanId, ProviderId, ProviderLifecycleAction, SessionId } from '@goodboy/types';
 import { CommandPalette } from './features/session/components/CommandPalette';
 import { BootSplash } from './app/components/BootSplash';
-import { ChatView } from './features/chat/components/ChatView';
-import { ContextPanel } from './features/context/components/ContextPanel';
+import { KeepAliveChatPanel, KeepAliveContextPanel } from './app/components/KeepAlivePanels';
+import { EmptyState } from './app/components/AppEmptyState';
 import { DeleteSessionDialog } from './features/session/components/DeleteSessionDialog';
 import { ArchiveSessionDialog } from './features/session/components/ArchiveSessionDialog';
 import { SettingsStudio } from './features/settings/components/SettingsStudio';
@@ -37,18 +37,18 @@ import {
 } from './features/permissions/components/DiffViewerDialog';
 import { ghCommitDiff } from './features/github/github';
 import { worktreeDiffCommit } from './features/worktree/worktree';
-import { DogMascot } from './shared/components/DogMascot';
 import { OnboardingCard } from './features/onboarding/OnboardingCard';
 import { OnboardingWizard } from './features/onboarding/OnboardingWizard';
 import { markStepComplete } from './features/onboarding/onboarding-store';
-import { BookOpen, MessageSquare, MessagesSquare } from 'lucide-react';
 import { useKeyboardShortcut } from './shared/hooks/useKeyboardShortcut';
 import { useProviderRefreshOnFocus } from './shared/hooks/useProviderRefreshOnFocus';
+import { useZoomShortcuts } from './shared/hooks/useZoomShortcuts';
+import { useEscapeToCloseDialog } from './shared/hooks/useEscapeToCloseDialog';
+import { useCommitLinkInterceptor } from './shared/hooks/useCommitLinkInterceptor';
 import {
   useAppStore,
   useCurrentSession,
   useCurrentWorkspace,
-  useSessionById,
   useSessions,
   useSessionSlots,
   useWorkspaces,
@@ -57,18 +57,8 @@ import { refreshPricingTable } from './features/providers/provider-pricing';
 import { useGithubPolling } from './features/github/hooks/useGithubPolling';
 import { useUpdaterPolling } from './features/updater/hooks/useUpdaterPolling';
 import { STORAGE_PREFIXES } from './shared/lib/storage-keys';
-import { openUrl } from './shared/lib/editor';
-import { applyStoredZoom, zoomIn, zoomOut, zoomReset } from './shared/lib/zoom';
 
 const CONTEXT_PANEL_KEY = (id: SessionId): string => `${STORAGE_PREFIXES.contextPanelOpen}${id}`;
-
-const ZOOM_ACTIONS: Record<string, () => Promise<void>> = {
-  '=': zoomIn,
-  '+': zoomIn,
-  '-': zoomOut,
-  _: zoomOut,
-  '0': zoomReset,
-};
 
 const KEEP_ALIVE_CAP = 5;
 
@@ -147,7 +137,7 @@ export const App = () => {
   const [budgetStudioScope, setBudgetStudioScope] = useState<BudgetScope | undefined>(undefined);
   const [workflowBuilderSessionId, setWorkflowBuilderSessionId] = useState<SessionId | null>(null);
   const [newSessionOpen, setNewSessionOpen] = useState(false);
-  const [commitDiff, setCommitDiff] = useState<{ repo: string; sha: string } | null>(null);
+  const { commitDiff, setCommitDiff } = useCommitLinkInterceptor();
   const [leftCollapsed, setLeftCollapsed] = useState(
     () =>
       typeof localStorage !== 'undefined' &&
@@ -169,28 +159,8 @@ export const App = () => {
   useProviderRefreshOnFocus();
   useUpdaterPolling();
   useWindowPresence();
-
-  useEffect(() => {
-    void applyStoredZoom();
-    const onShortcut = (e: KeyboardEvent) => {
-      if (!(e.metaKey || e.ctrlKey)) {
-        return;
-      }
-      if (e.key === 'r' || e.key === 'R') {
-        e.preventDefault();
-        window.location.reload();
-        return;
-      }
-      const action = ZOOM_ACTIONS[e.key];
-      if (!action) {
-        return;
-      }
-      e.preventDefault();
-      void action();
-    };
-    window.addEventListener('keydown', onShortcut);
-    return () => window.removeEventListener('keydown', onShortcut);
-  }, []);
+  useZoomShortcuts();
+  useEscapeToCloseDialog();
 
   useEffect(() => {
     const handler = (event: Event) => {
@@ -448,45 +418,6 @@ export const App = () => {
     const handler = () => setSwitcherOpen(true);
     window.addEventListener('goodboy:open-workspace-switcher', handler);
     return () => window.removeEventListener('goodboy:open-workspace-switcher', handler);
-  }, []);
-
-  useEffect(() => {
-    const COMMIT_RE = /^https?:\/\/github\.com\/([^/]+\/[^/]+)\/commit\/([0-9a-f]{7,40})/i;
-    const onClick = (e: MouseEvent) => {
-      if (e.defaultPrevented || e.button !== 0 || e.metaKey || e.ctrlKey) {
-        return;
-      }
-      const anchor = (e.target as HTMLElement | null)?.closest?.('a');
-      const href = anchor?.getAttribute('href');
-      if (!href) {
-        return;
-      }
-      const commit = href.match(COMMIT_RE);
-      if (commit) {
-        e.preventDefault();
-        setCommitDiff({ repo: commit[1] as string, sha: commit[2] as string });
-        return;
-      }
-      if (/^(https?:|mailto:)/i.test(href)) {
-        e.preventDefault();
-        void openUrl(href);
-      }
-    };
-    document.addEventListener('click', onClick);
-    return () => document.removeEventListener('click', onClick);
-  }, []);
-
-  useEffect(() => {
-    const onEsc = (e: KeyboardEvent) => {
-      if (e.key !== 'Escape') {
-        return;
-      }
-      e.preventDefault();
-      const dialogs = document.querySelectorAll<HTMLDialogElement>('dialog[open]');
-      dialogs[dialogs.length - 1]?.close();
-    };
-    document.addEventListener('keydown', onEsc, { capture: true });
-    return () => document.removeEventListener('keydown', onEsc, { capture: true });
   }, []);
 
   useEffect(() => {
@@ -1006,211 +937,3 @@ export const App = () => {
     </ToastProvider>
   );
 };
-
-type KeepAliveChatPanelProps = {
-  readonly sessionId: SessionId;
-  readonly isActive: boolean;
-};
-
-function KeepAliveChatPanel({ sessionId, isActive }: KeepAliveChatPanelProps) {
-  const session = useSessionById(sessionId);
-  if (!session) {
-    return null;
-  }
-  return (
-    <div hidden={!isActive} className="absolute inset-0">
-      <ChatView session={session} isActive={isActive} />
-    </div>
-  );
-}
-
-type KeepAliveContextPanelProps = {
-  readonly sessionId: SessionId;
-  readonly isActive: boolean;
-  readonly collapsed: boolean;
-  readonly onCollapse: () => void;
-  readonly onExpand: () => void;
-};
-
-function KeepAliveContextPanel({
-  sessionId,
-  isActive,
-  collapsed,
-  onCollapse,
-  onExpand,
-}: KeepAliveContextPanelProps) {
-  const session = useSessionById(sessionId);
-  if (!session) {
-    return null;
-  }
-  return (
-    <div hidden={!isActive} className="absolute inset-0">
-      <ContextPanel
-        session={session}
-        isActive={isActive}
-        collapsed={collapsed}
-        onCollapse={onCollapse}
-        onExpand={onExpand}
-      />
-    </div>
-  );
-}
-
-function EmptyState({
-  hasWorkspace,
-  onAddWorkspace,
-}: {
-  hasWorkspace: boolean;
-  onAddWorkspace: () => void;
-}) {
-  if (!hasWorkspace) {
-    return <OnboardingScreen onAddWorkspace={onAddWorkspace} />;
-  }
-
-  return (
-    <div className="relative flex h-full w-full items-center justify-center overflow-hidden px-6">
-      <div
-        className="pointer-events-none absolute inset-0"
-        style={{
-          background:
-            'radial-gradient(ellipse at center, transparent 0%, transparent 40%, var(--color-background) 100%)',
-        }}
-        aria-hidden
-      />
-      <div className="relative flex max-w-md flex-col items-center gap-6 text-center">
-        <EmptyStateLogo />
-        <div className="flex flex-col gap-2.5">
-          <h2 className="text-2xl font-semibold tracking-tight text-foreground">
-            Pick up where you left off
-          </h2>
-          <p className="max-w-sm text-sm leading-relaxed text-muted-foreground">
-            Create a new session from the sidebar, or jump back into an existing one. Each session
-            lives in its own worktree.
-          </p>
-        </div>
-        <KeyboardHints hasWorkspace />
-      </div>
-    </div>
-  );
-}
-
-function OnboardingScreen({ onAddWorkspace }: { onAddWorkspace: () => void }) {
-  return (
-    <div className="relative flex h-full w-full items-center justify-center overflow-hidden px-6">
-      <div
-        className="pointer-events-none absolute inset-0"
-        style={{
-          background:
-            'radial-gradient(ellipse at center, transparent 0%, transparent 40%, var(--color-background) 100%)',
-        }}
-        aria-hidden
-      />
-
-      <div className="relative flex max-w-2xl flex-col items-center gap-10 text-center">
-        <EmptyStateLogo />
-
-        <div className="flex flex-col gap-3">
-          <h2 className="text-2xl font-semibold tracking-tight text-foreground">
-            Welcome to Goodboy
-          </h2>
-          <p className="max-w-md text-sm leading-relaxed text-muted-foreground">
-            Point at a git repo to create your first workspace. Every session spins up its own
-            worktree and branch, your main checkout stays untouched.
-          </p>
-        </div>
-
-        <button
-          type="button"
-          onClick={onAddWorkspace}
-          className="inline-flex items-center gap-2 rounded-lg bg-primary px-6 py-2.5 text-sm font-medium text-primary-foreground shadow-sm transition-colors hover:bg-primary/90"
-        >
-          Add workspace
-        </button>
-
-        <AppLayoutPreview />
-      </div>
-    </div>
-  );
-}
-
-function AppLayoutPreview() {
-  return (
-    <div className="flex w-full max-w-2xl gap-3">
-      <div className="flex w-[30%] flex-col items-center gap-3 rounded-lg border border-border-soft/30 bg-subtle/15 px-5 py-6">
-        <div className="flex size-10 items-center justify-center rounded-lg bg-muted/40">
-          <MessagesSquare size={20} className="text-muted-foreground/60" aria-hidden />
-        </div>
-        <span className="text-xs font-semibold text-muted-foreground">Sessions</span>
-        <p className="text-2xs leading-relaxed text-muted-foreground/50">
-          Switch workspaces, manage sessions, track agents and workflow progress.
-        </p>
-      </div>
-
-      <div className="flex flex-1 flex-col items-center gap-3 rounded-lg border border-border-soft/30 bg-background/30 px-6 py-6">
-        <div className="flex size-10 items-center justify-center rounded-lg bg-primary/10">
-          <MessageSquare size={20} className="text-primary/60" aria-hidden />
-        </div>
-        <span className="text-xs font-semibold text-muted-foreground">Chat</span>
-        <p className="text-2xs leading-relaxed text-muted-foreground/50">
-          Talk to your agents, send instructions, and watch execution unfold in real time.
-        </p>
-      </div>
-
-      <div className="flex w-[26%] flex-col items-center gap-3 rounded-lg border border-border-soft/30 bg-subtle/15 px-5 py-6">
-        <div className="flex size-10 items-center justify-center rounded-lg bg-muted/40">
-          <BookOpen size={20} className="text-muted-foreground/60" aria-hidden />
-        </div>
-        <span className="text-xs font-semibold text-muted-foreground">Context</span>
-        <p className="text-2xs leading-relaxed text-muted-foreground/50">
-          Inject context slots, review touched files, and check PR details at a glance.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-function EmptyStateLogo() {
-  return (
-    <div className="relative">
-      <div className="absolute -inset-6 rounded-full bg-primary/10 blur-2xl" />
-      <div className="relative flex h-24 w-24 items-center justify-center rounded-lg border border-border-soft/40 bg-subtle/40 shadow-lg backdrop-blur-sm">
-        <DogMascot size={56} className="text-foreground" />
-      </div>
-    </div>
-  );
-}
-
-function KeyboardHints({ hasWorkspace }: { hasWorkspace: boolean }) {
-  if (!hasWorkspace) {
-    return null;
-  }
-  return (
-    <div className="mt-2 flex flex-wrap items-center justify-center gap-2 text-2xs text-muted-foreground">
-      <span className="inline-flex items-center gap-1">
-        <Kbd>⌘</Kbd>
-        <Kbd>K</Kbd>
-        <span className="ml-1">command palette</span>
-      </span>
-      <span className="text-muted-foreground/30">·</span>
-      <span className="inline-flex items-center gap-1">
-        <Kbd>⌘</Kbd>
-        <Kbd>,</Kbd>
-        <span className="ml-1">settings</span>
-      </span>
-      <span className="text-muted-foreground/30">·</span>
-      <span className="inline-flex items-center gap-1">
-        <Kbd>⌘</Kbd>
-        <Kbd>/</Kbd>
-        <span className="ml-1">shortcuts</span>
-      </span>
-    </div>
-  );
-}
-
-function Kbd({ children }: { children: React.ReactNode }) {
-  return (
-    <kbd className="inline-flex h-4 min-w-[1rem] items-center justify-center rounded border border-border-soft bg-subtle/60 px-1 font-mono text-[10px] font-medium text-muted-foreground">
-      {children}
-    </kbd>
-  );
-}
